:recycle: refactor: 컴포넌트를 폴더 안에 넣는다
